+++ src/PhpPresentation/Writer/PowerPoint2007.php
@@ -17,23 +17,15 @@
 
 namespace PhpOffice\PhpPresentation\Writer;
 
+use DirectoryIterator;
 use PhpOffice\PhpPresentation\HashTable;
 use PhpOffice\PhpPresentation\PhpPresentation;
-use PhpOffice\PhpPresentation\Shape\Drawing as DrawingShape;
-use PhpOffice\PhpPresentation\Shape\MemoryDrawing as MemoryDrawingShape;
+use PhpOffice\PhpPresentation\Shape\AbstractDrawing;
 use PhpOffice\PhpPresentation\Shape\Chart as ChartShape;
-use PhpOffice\PhpPresentation\Slide\Background\Image;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractLayoutPack;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Chart;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\ContentTypes;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\DocProps;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Drawing;
+use PhpOffice\PhpPresentation\Shape\Group;
+use PhpOffice\PhpPresentation\Shape\Table;
+use PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack\AbstractLayoutPack;
 use PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack\PackDefault;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\PptProps;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Presentation;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Rels;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Slide;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007\Theme;
 
 /**
  * \PhpOffice\PhpPresentation\Writer\PowerPoint2007
@@ -46,13 +38,6 @@ class PowerPoint2007 implements WriterInterface
      * @var boolean
      */
     protected $office2003comp = false;
-
-    /**
-     * Private writer parts
-     *
-     * @var \PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractPart[]
-     */
-    protected $writerParts;
 
     /**
      * Private PhpPresentation
@@ -85,7 +70,7 @@ class PowerPoint2007 implements WriterInterface
     /**
      * Layout pack to use
      *
-     * @var \PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractLayoutPack
+     * @var \PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack\AbstractLayoutPack
      */
     protected $layoutPack;
 
@@ -105,39 +90,8 @@ class PowerPoint2007 implements WriterInterface
         // Set layout pack
         $this->layoutPack = new PackDefault();
 
-        // Initialise writer parts
-        $this->writerParts['contenttypes'] = new ContentTypes();
-        $this->writerParts['pptprops']     = new PptProps();
-        $this->writerParts['docprops']     = new DocProps();
-        $this->writerParts['rels']         = new Rels();
-        $this->writerParts['theme']        = new Theme();
-        $this->writerParts['presentation'] = new Presentation();
-        $this->writerParts['slide']        = new Slide();
-        $this->writerParts['drawing']      = new Drawing();
-        $this->writerParts['chart']        = new Chart();
-
-        // Assign parent WriterInterface
-        foreach ($this->writerParts as $writer) {
-            $writer->setParentWriter($this);
-        }
-
         // Set HashTable variables
         $this->drawingHashTable = new HashTable();
-    }
-
-    /**
-     * Get writer part
-     *
-     * @param  string                                         $pPartName Writer part name
-     * @return \PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractPart
-     */
-    public function getWriterPart($pPartName = '')
-    {
-        if ($pPartName != '' && isset($this->writerParts[strtolower($pPartName)])) {
-            return $this->writerParts[strtolower($pPartName)];
-        } else {
-            return null;
-        }
     }
 
     /**
@@ -161,184 +115,39 @@ class PowerPoint2007 implements WriterInterface
                 }
             }
             
-            $wPartDrawing = $this->getWriterPart('Drawing');
-            if (!$wPartDrawing instanceof Drawing) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\Drawing');
-            }
-            $wPartContentTypes = $this->getWriterPart('ContentTypes');
-            if (!$wPartContentTypes instanceof ContentTypes) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\ContentTypes');
-            }
-            $wPartRels = $this->getWriterPart('Rels');
-            if (!$wPartRels instanceof Rels) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\Rels');
-            }
-            $wPartDocProps = $this->getWriterPart('DocProps');
-            if (!$wPartDocProps instanceof DocProps) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\DocProps');
-            }
-            $wPartTheme = $this->getWriterPart('Theme');
-            if (!$wPartTheme instanceof Theme) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\Theme');
-            }
-            $wPartPresentation = $this->getWriterPart('Presentation');
-            if (!$wPartPresentation instanceof Presentation) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\Presentation');
-            }
-            $wPartSlide = $this->getWriterPart('Slide');
-            if (!$wPartSlide instanceof Slide) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\Slide');
-            }
-            $wPartChart = $this->getWriterPart('Chart');
-            if (!$wPartChart instanceof Chart) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\Chart');
-            }
-            $wPptProps = $this->getWriterPart('PptProps');
-            if (!$wPptProps instanceof PptProps) {
-                throw new \Exception('The $parentWriter is not an instance of \PhpOffice\PhpPresentation\Writer\PowerPoint2007\PptProps');
-            }
-            
             // Create drawing dictionary
-            $this->drawingHashTable->addFromSource($wPartDrawing->allDrawings($this->presentation));
+            $this->drawingHashTable->addFromSource($this->allDrawings());
 
-            // Create new ZIP file and open it for writing
-            $objZip = new \ZipArchive();
+            $oZip = new \ZipArchive();
 
             // Try opening the ZIP file
-            if ($objZip->open($pFilename, \ZipArchive::OVERWRITE) !== true) {
-                if ($objZip->open($pFilename, \ZipArchive::CREATE) !== true) {
+            if ($oZip->open($pFilename, \ZipArchive::OVERWRITE) !== true) {
+                if ($oZip->open($pFilename, \ZipArchive::CREATE) !== true) {
                     throw new \Exception("Could not open " . $pFilename . " for writing.");
                 }
             }
 
-            // Add [Content_Types].xml to ZIP file
-            $objZip->addFromString('[Content_Types].xml', $wPartContentTypes->writeContentTypes($this->presentation));
-
-            // Add PPT properties and styles to ZIP file - Required for Apple Keynote compatibility.
-            $objZip->addFromString('ppt/presProps.xml', $wPptProps->writePresProps($this->presentation));
-            $objZip->addFromString('ppt/tableStyles.xml', $wPptProps->writeTableStyles());
-            $objZip->addFromString('ppt/viewProps.xml', $wPptProps->writeViewProps($this->presentation));
-
-            // Add relationships to ZIP file
-            $objZip->addFromString('_rels/.rels', $wPartRels->writeRelationships($this->presentation));
-            $objZip->addFromString('ppt/_rels/presentation.xml.rels', $wPartRels->writePresentationRelationships($this->presentation));
-
-            // Add document properties to ZIP file
-            $objZip->addFromString('docProps/app.xml', $wPartDocProps->writeDocPropsApp($this->presentation));
-            $objZip->addFromString('docProps/core.xml', $wPartDocProps->writeDocPropsCore($this->presentation));
-            $objZip->addFromString('docProps/custom.xml', $wPartDocProps->writeDocPropsCustom($this->presentation));
-            // Add Thumbnail to ZIP file
-            if ($this->presentation->getPresentationProperties()->getThumbnailPath()) {
-                $pathThumbnail = file_get_contents($this->presentation->getPresentationProperties()->getThumbnailPath());
-                $gdImage = imagecreatefromstring($pathThumbnail);
-                if ($gdImage) {
-                    ob_start();
-                    imagejpeg($gdImage);
-                    $imageContents = ob_get_contents();
-                    ob_end_clean();
-                    imagedestroy($gdImage);
-
-                    $objZip->addFromString('docProps/thumbnail.jpeg', $imageContents);
+            $oDir = new DirectoryIterator(dirname(__FILE__).'\PowerPoint2007');
+            foreach ($oDir as $oFile) {
+                if (!$oFile->isFile()) {
+                    continue;
                 }
-            }
+                $class = __NAMESPACE__.'\\PowerPoint2007\\'.$oFile->getBasename('.php');
+                $o = new \ReflectionClass($class);
 
-            $masterSlides = $this->getLayoutPack()->getMasterSlides();
-            foreach ($masterSlides as $masterSlide) {
-                // Add themes to ZIP file
-                $objZip->addFromString('ppt/theme/_rels/theme' . $masterSlide['masterid'] . '.xml.rels', $wPartRels->writeThemeRelationships($masterSlide['masterid']));
-                $objZip->addFromString('ppt/theme/theme' . $masterSlide['masterid'] . '.xml', $wPartTheme->writeTheme($masterSlide['masterid']));
-                // Add slide masters to ZIP file
-                $objZip->addFromString('ppt/slideMasters/_rels/slideMaster' . $masterSlide['masterid'] . '.xml.rels', $wPartRels->writeSlideMasterRelationships($masterSlide['masterid']));
-                $objZip->addFromString('ppt/slideMasters/slideMaster' . $masterSlide['masterid'] . '.xml', $masterSlide['body']);
-            }
-
-            // Add slide layouts to ZIP file
-            $slideLayouts = $this->getLayoutPack()->getLayouts();
-            foreach ($slideLayouts as $key => $layout) {
-                $objZip->addFromString('ppt/slideLayouts/_rels/slideLayout' . $key . '.xml.rels', $wPartRels->writeSlideLayoutRelationships($key, $layout['masterid']));
-                $objZip->addFromString('ppt/slideLayouts/slideLayout' . $key . '.xml', utf8_encode($layout['body']));
-            }
-
-            // Add layoutpack relations
-            $otherRelations = $this->getLayoutPack()->getMasterSlideRelations();
-            foreach ($otherRelations as $otherRelation) {
-                if (strpos($otherRelation['target'], 'http://') !== 0) {
-                    $objZip->addFromString($this->absoluteZipPath('ppt/slideMasters/' . $otherRelation['target']), $otherRelation['contents']);
+                if ($o->isAbstract() || !$o->isSubclassOf('PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractDecoratorWriter')){
+                    continue;
                 }
-            }
-            $otherRelations = $this->getLayoutPack()->getThemeRelations();
-            foreach ($otherRelations as $otherRelation) {
-                if (strpos($otherRelation['target'], 'http://') !== 0) {
-                    $objZip->addFromString($this->absoluteZipPath('ppt/theme/' . $otherRelation['target']), $otherRelation['contents']);
-                }
-            }
-            $otherRelations = $this->getLayoutPack()->getLayoutRelations();
-            foreach ($otherRelations as $otherRelation) {
-                if (strpos($otherRelation['target'], 'http://') !== 0) {
-                    $objZip->addFromString($this->absoluteZipPath('ppt/slideLayouts/' . $otherRelation['target']), $otherRelation['contents']);
-                }
-            }
-
-            // Add presentation to ZIP file
-            $objZip->addFromString('ppt/presentation.xml', $wPartPresentation->writePresentation($this->presentation));
-
-            // Add slides (drawings, ...) and slide relationships (drawings, ...)
-            for ($i = 0; $i < $this->presentation->getSlideCount(); ++$i) {
-                $oSlide = $this->presentation->getSlide($i);
-                // Add slide
-                $objZip->addFromString('ppt/slides/_rels/slide' . ($i + 1) . '.xml.rels', $wPartRels->writeSlideRelationships($oSlide));
-                $objZip->addFromString('ppt/slides/slide' . ($i + 1) . '.xml', $wPartSlide->writeSlide($oSlide));
-                // Add note slide
-                if ($oSlide->getNote()->getShapeCollection()->count() > 0) {
-                    $objZip->addFromString('ppt/notesSlides/notesSlide' . ($i + 1) . '.xml', $wPartSlide->writeNote($oSlide->getNote()));
-                }
-                // Add background image slide
-                $oBkgImage = $oSlide->getBackground();
-                if ($oBkgImage instanceof Image) {
-                    $objZip->addFromString('ppt/media/'.$oBkgImage->getIndexedFilename($i), file_get_contents($oBkgImage->getPath()));
-                }
-            }
-
-            // Add media
-            for ($i = 0; $i < $this->getDrawingHashTable()->count(); ++$i) {
-                $shape = $this->getDrawingHashTable()->getByIndex($i);
-                if ($shape instanceof DrawingShape) {
-                    $imagePath     = $shape->getPath();
-
-                    if (strpos($imagePath, 'zip://') !== false) {
-                        $imagePath         = substr($imagePath, 6);
-                        $imagePathSplitted = explode('#', $imagePath);
-
-                        $imageZip = new \ZipArchive();
-                        $imageZip->open($imagePathSplitted[0]);
-                        $imageContents = $imageZip->getFromName($imagePathSplitted[1]);
-                        $imageZip->close();
-                        unset($imageZip);
-                    } else {
-                        $imageContents = file_get_contents($imagePath);
-                    }
-
-                    $objZip->addFromString('ppt/media/' . str_replace(' ', '_', $shape->getIndexedFilename()), $imageContents);
-                } elseif ($shape instanceof MemoryDrawingShape) {
-                    ob_start();
-                    call_user_func($shape->getRenderingFunction(), $shape->getImageResource());
-                    $imageContents = ob_get_contents();
-                    ob_end_clean();
-
-                    $objZip->addFromString('ppt/media/' . str_replace(' ', '_', $shape->getIndexedFilename()), $imageContents);
-                } elseif ($shape instanceof ChartShape) {
-                    $objZip->addFromString('ppt/charts/' . $shape->getIndexedFilename(), $wPartChart->writeChart($shape));
-
-                    // Chart relations?
-                    if ($shape->hasIncludedSpreadsheet()) {
-                        $objZip->addFromString('ppt/charts/_rels/' . $shape->getIndexedFilename() . '.rels', $wPartRels->writeChartRelationships($shape));
-                        $objZip->addFromString('ppt/embeddings/' . $shape->getIndexedFilename() . '.xlsx', $wPartChart->writeSpreadsheet($this->presentation, $shape, $pFilename . '.xlsx'));
-                    }
-                }
+                $oService = $o->newInstance();
+                $oService->setZip($oZip);
+                $oService->setPresentation($this->presentation);
+                $oService->setDrawingHashTable($this->drawingHashTable);
+                $oZip = $oService->render();
+                unset($oService);
             }
 
             // Close file
-            if ($objZip->close() === false) {
+            if ($oZip->close() === false) {
                 throw new \Exception("Could not close zip file $pFilename.");
             }
 
@@ -464,7 +273,7 @@ class PowerPoint2007 implements WriterInterface
     /**
      * Get layout pack to use
      *
-     * @return \PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractLayoutPack
+     * @return \PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack\AbstractLayoutPack
      */
     public function getLayoutPack()
     {
@@ -474,7 +283,7 @@ class PowerPoint2007 implements WriterInterface
     /**
      * Set layout pack to use
      *
-     * @param \PhpOffice\PhpPresentation\Writer\PowerPoint2007\AbstractLayoutPack $pValue
+     * @param \PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack\AbstractLayoutPack $pValue
      * @return \PhpOffice\PhpPresentation\Writer\PowerPoint2007
      */
     public function setLayoutPack(AbstractLayoutPack $pValue = null)
@@ -485,30 +294,40 @@ class PowerPoint2007 implements WriterInterface
     }
 
     /**
-     * Determine absolute zip path
+     * Get an array of all drawings
      *
-     * @param  string $path
-     * @return string
+     * @return \PhpOffice\PhpPresentation\Shape\AbstractDrawing[] All drawings in PhpPresentation
+     * @throws \Exception
      */
-    protected function absoluteZipPath($path)
+    protected function allDrawings()
     {
-        $path      = str_replace(array(
-            '/',
-            '\\'
-        ), DIRECTORY_SEPARATOR, $path);
-        $parts     = array_filter(explode(DIRECTORY_SEPARATOR, $path), 'strlen');
-        $absolutes = array();
-        foreach ($parts as $part) {
-            if ('.' == $part) {
+        // Get an array of all drawings
+        $aDrawings  = array();
+
+        // Loop trough PhpPresentation
+        foreach ($this->getPhpPresentation()->getAllSlides() as $oSlide) {
+            $oCollection = $oSlide->getShapeCollection();
+            if ($oCollection->count() <= 0) {
                 continue;
             }
-            if ('..' == $part) {
-                array_pop($absolutes);
-            } else {
-                $absolutes[] = $part;
+            $oIterator = $oCollection->getIterator();
+            while ($oIterator->valid()) {
+                if ($oIterator->current() instanceof AbstractDrawing && !($oIterator->current() instanceof Table)) {
+                    $aDrawings[] = $oIterator->current();
+                } elseif ($oIterator->current() instanceof Group) {
+                    $oSubIterator = $oIterator->current()->getShapeCollection()->getIterator();
+                    while ($oSubIterator->valid()) {
+                        if ($oSubIterator->current() instanceof AbstractDrawing && !($oSubIterator->current() instanceof Table)) {
+                            $aDrawings[] = $oSubIterator->current();
+                        }
+                        $oSubIterator->next();
+                    }
+                }
+
+                $oIterator->next();
             }
         }
 
-        return implode('/', $absolutes);
+        return $aDrawings;
     }
 }